add listing ownership & authorization

+++ app/Http/Controllers/ListingController.php
@@ -45,6 +45,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         // Session::flash('message', 'listing created');
@@ -55,10 +57,17 @@ class ListingController extends Controller
 
     // show edit form
     public function edit(Listing $listing){
+        if($listing->user_id != auth()->id()){
+            abort(403, 'forbidden');
+        }
         return view('listings.edit', ['listing' => $listing]);
     }
 
     public function update(Request $request, Listing $listing){
+        if($listing->user_id != auth()->id()){
+            abort(403, 'forbidden');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -79,7 +88,14 @@ class ListingController extends Controller
     }
 
     public function destroy(Listing $listing){
+        if($listing->user_id != auth()->id()){
+            abort(403, 'forbidden');
+        }
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully');
+    }
+
+    public function manage(){
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
